renamed license validator to more info
