+++ app/templates/more-info.hbs
@@ -3,7 +3,7 @@
   <p>
     We'll gather more info from the user about license/attribution/update frequency here.
   </p>
-  {{#with (changeset model LicenseValidator) as |changeset|}}
+  {{#with (changeset model InfoValidator) as |changeset|}}
     <div class="form-content">
       <div class="ui tiny message">
         <p> <span class="required"> * </span>
